new product as noshmishmosh
The Nosh Mish Mosh is a recipe and ingredient meal delivery service. We ship the raw materials and you get to cook it at your home! We’ve decided to hire a data analyst to help them make product and interface decisions. Get started and we’ll be able to judge the amount of data we’ll need for these differences to be meaningful.